feat(test): move them close to the source code
